Fixes for APPM-1365

+++ components/org.wso2.carbon.appmgt.impl/src/main/java/org/wso2/carbon/appmgt/impl/APIProviderImpl.java
@@ -104,11 +104,19 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
     /**
      * Delete business owner.
      * @param businessOwnerId ID of the owner.
+     * @return
      * @throws AppManagementException
      */
     @Override
-    public void deleteBusinessOwner(String businessOwnerId) throws AppManagementException{
-         appMDAO.deleteBusinessOwner(businessOwnerId);
+    public boolean deleteBusinessOwner(String businessOwnerId) throws AppManagementException{
+        boolean isAssociatedWithApps = appMDAO.isAssociatedWithApps(businessOwnerId, registry,tenantDomain);
+        if (!isAssociatedWithApps) {
+            appMDAO.deleteBusinessOwner(businessOwnerId);
+            // return true if business owner is successfully deleted.
+            return true;
+        }
+        // return false if business owner is associated with one or more web apps.
+        return false;
     }
     /**
      *Update a business owner.
@@ -141,7 +149,7 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
      */
     @Override
     public List<BusinessOwner> getBusinessOwners() throws AppManagementException {
-        return appMDAO.getBusinessOwners();
+        return appMDAO.getBusinessOwners(tenantId);
     }
 
     /**
@@ -166,12 +174,12 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
     @Override
     public  List<BusinessOwner> searchBusinessOwners(int startIndex, int pageSize, String searchValue) throws
                                                                                           AppManagementException {
-        return appMDAO.searchBusinessOwners(startIndex, pageSize, searchValue);
+        return appMDAO.searchBusinessOwners(startIndex, pageSize, searchValue, tenantId);
     }
 
     @Override
     public  int getBusinessOwnersCount() throws AppManagementException {
-        return appMDAO.getBusinessOwnersCount();
+        return appMDAO.getBusinessOwnersCount(tenantId);
     }
 
     /**
@@ -181,7 +189,7 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
      */
     @Override
     public int saveBusinessOwner(BusinessOwner businessOwner) throws AppManagementException {
-        return appMDAO.saveBusinessOwner(businessOwner);
+        return appMDAO.saveBusinessOwner(businessOwner, tenantId);
     }
 
     /**
